// src/generated/api/models/ProfileImageResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProfileImageResponse
 */
export interface ProfileImageResponse {
    /**
     * 
     * @type {number}
     * @memberof ProfileImageResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ProfileImageResponse
     */
    imageUrl?: string;
}

/**
 * Check if a given object implements the ProfileImageResponse interface.
 */
export function instanceOfProfileImageResponse(value: object): value is ProfileImageResponse {
    return true;
}

export function ProfileImageResponseFromJSON(json: any): ProfileImageResponse {
    return ProfileImageResponseFromJSONTyped(json, false);
}

export function ProfileImageResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfileImageResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
    };
}

export function ProfileImageResponseToJSON(json: any): ProfileImageResponse {
    return ProfileImageResponseToJSONTyped(json, false);
}

export function ProfileImageResponseToJSONTyped(value?: ProfileImageResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'imageUrl': value['imageUrl'],
    };
}