// src/generated/api/models/UpdateProfileRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface UpdateProfileRequest
 */
export interface UpdateProfileRequest {
  /**
   * 이메일
   * @type {string}
   * @memberof UpdateProfileRequest
   */
  email: string;
  /**
   * 학번
   * @type {string}
   * @memberof UpdateProfileRequest
   */
  studentNumber: string;
  /**
   * 감자 동아리 기수
   * @type {number}
   * @memberof UpdateProfileRequest
   */
  gamjaBatch: number;
  /**
   * 포지션/직책
   * @type {string}
   * @memberof UpdateProfileRequest
   */
  position: UpdateProfileRequestPositionEnum;
}

/**
 * @export
 */
export const UpdateProfileRequestPositionEnum = {
  Backend: "BACKEND",
  Frontend: "FRONTEND",
  Designer: "DESIGNER",
  AppDeveloper: "APP_DEVELOPER",
} as const;
export type UpdateProfileRequestPositionEnum =
  (typeof UpdateProfileRequestPositionEnum)[keyof typeof UpdateProfileRequestPositionEnum];

/**
 * Check if a given object implements the UpdateProfileRequest interface.
 */
export function instanceOfUpdateProfileRequest(value: object): value is UpdateProfileRequest {
  if (!("email" in value) || value["email"] === undefined) return false;
  if (!("studentNumber" in value) || value["studentNumber"] === undefined) return false;
  if (!("gamjaBatch" in value) || value["gamjaBatch"] === undefined) return false;
  if (!("position" in value) || value["position"] === undefined) return false;
  return true;
}

export function UpdateProfileRequestFromJSON(json: any): UpdateProfileRequest {
  return UpdateProfileRequestFromJSONTyped(json, false);
}

export function UpdateProfileRequestFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): UpdateProfileRequest {
  if (json == null) {
    return json;
  }
  return {
    email: json["email"],
    studentNumber: json["studentNumber"],
    gamjaBatch: json["gamjaBatch"],
    position: json["position"],
  };
}

export function UpdateProfileRequestToJSON(json: any): UpdateProfileRequest {
  return UpdateProfileRequestToJSONTyped(json, false);
}

export function UpdateProfileRequestToJSONTyped(
  value?: UpdateProfileRequest | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    email: value["email"],
    studentNumber: value["studentNumber"],
    gamjaBatch: value["gamjaBatch"],
    position: value["position"],
  };
}